USER_STORY_AGENT/requirements.txt
anthropic
openpyxl
pandas
PyPDF2
python-docx
python-dotenv
streamlit
mcp
pytesseract
Pillow
beautifulsoup4
lxml
